requirements.txt
click>=8.0.3
google_api_python_client==2.99.0
google-cloud-bigquery==3.13.0
google_auth_oauthlib==1.1.0
pandas==2.1.0
protobuf>=3.12.4
slack_sdk==3.23.0
resotoclient==1.5.0
matplotlib>=3.8